% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{credible_interval}
\alias{credible_interval}
\title{Compute upper and lower values of the 95\% credible interval}
\usage{
credible_interval(x)
}
\arguments{
\item{x}{posterior values for a parameter}
}
\value{
numeric vector. Element 1 is the lower 95\% CI, element 2 is the upper 95\% CI.
}
\description{
Compute upper and lower values of the 95\% credible interval
}
