% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scenarios.R
\name{generate_scenarios}
\alias{generate_scenarios}
\title{Generate parameter settings for each simulation iteration}
\usage{
generate_scenarios(seed = 3547912)
}
\arguments{
\item{seed}{random seed to use for generating parameter settings for each iteration. Defaults to 3547912, which is the seed I use for my thesis.}
}
\value{
data frame with 36.000 rows and 11 columns
\describe{
   \item{n}{int. number of subjects}
   \item{n_t}{int. number of observed data points for each subject}
   \item{zeta}{float. between-subject variance for the means of the emission distributions}
   \item{Q}{float. between-subject variance for the intercepts of the transition-probability matrix}
   \item{scenario_id}{string .unique id of the simulation scenario}
   \item{rank}{int. iteration number of the simulation scenario}
   \item{iteration_id}{string. unique id of the iteration}
   \item{dsim_seed}{int. random seed used to generate the data}
   \item{model_seed}{int. random seed used to run the mHMM}
   \item{start_gamma}{json. initial values for the between-subject transition probability matrix}
   \item{start_emiss}{json. intiial values for each of the 3 emission distributions}
}
}
\description{
Generate parameter settings for each simulation iteration
}
\details{
In my simulation study, I test 250 iterations for each of 144 different scenarios. Hence, this function
         returns a data frame with 36.000 rows - one for each iteration - that contains specific information
         used to generate and model the data
}
