% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{percent_bias}
\alias{percent_bias}
\title{Compute the parameter bias of a vector estimated parameters versus the ground-truth value of that parameter}
\usage{
percent_bias(true_param_value, simulated_param_values)
}
\arguments{
\item{true_param_value}{numeric. Value of the ground-truth parameter.}

\item{simulated_param_values}{k-length numeric vector. k >= 1 and holds the parameter values of the estimated parameters.}
}
\value{
percentage
}
\description{
Compute the parameter bias of a vector estimated parameters versus the ground-truth value of that parameter
}
\details{
This function computes the percentage bias by first squaring differences, then taking the square root. This way, positive and negative values are not cancelled out.
}
