% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot histograms of between-subject means}
\usage{
plot_posterior(x, ...)
}
\arguments{
\item{x}{an mHMM_cont object}

\item{param}{string. Which of the parameters should be plotted? Must be one of 'emiss_mu_bar', 'gamma_int_bar', 'emiss_var_bar' or 'emiss_varmu_bar'}

\item{var}{int. Which emission distribution should be plotted? Must be an integer equal to or less than the total number of emission distributions. Index of variables is the same as the order in which they appear in the input data given to the mHMM.}

\item{ground_truth}{numeric. Vector of ground-truth values for the parameters equal to the number of emission distributions. Will be added as a red dotted line.}
}
\value{
plots a histogram of the parameter of interest for each of the hypothesized latent states
}
\description{
Plot histograms of between-subject means
}
