% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_posterior_means}
\alias{plot_posterior_means}
\title{Plot histograms of between-subject means}
\usage{
plot_posterior_means(x, ...)
}
\arguments{
\item{x}{an mHMM_cont object}
}
\value{
plots the between-subject means
}
\description{
Plot histograms of between-subject means
}
