% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_simulation_settings}
\alias{query_simulation_settings}
\title{Fetch simulation settings used for this simulation scenario}
\usage{
query_simulation_settings()
}
\description{
Fetch simulation settings used for this simulation scenario
}
