% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{register_simulation_outcomes}
\alias{register_simulation_outcomes}
\title{Send the results of the simulation to the master node}
\usage{
register_simulation_outcomes(uid, status, diagnostics)
}
\arguments{
\item{uid}{Character. unique ID of this simulation program}

\item{status}{Numeric. http status of the simulation program. See <https://www.restapitutorial.com/httpstatuscodes.html>.}

\item{diagnostics}{List. diagnostic information on the simulation scenario.}
}
\value{

}
\description{
Send the results of the simulation to the master node
}
