% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_mHMM}
\alias{run_mHMM}
\title{Run an mHMM on a simulated sleep dataset}
\usage{
run_mHMM(
  data,
  start_values,
  model_seed,
  mcmc_iterations = 2000,
  mcmc_burn_in = 1000,
  show_progress = TRUE
)
}
\arguments{
\item{data}{Matrix. data set used to run the mHMM. See s_data parameter in \link[mHMMbayes]{mHMM_cont}.}

\item{start_values}{List. start values for relevant parameters. See start_val parameter in \link[mHMMbayes]{mHMM_cont}.}

\item{model_seed}{Int. Random seed that is set before running the model.}

\item{mcmc_iterations}{Int. number of iterations for the MCMC algorithm. Defaults to 1000. See mcmc parameter in \link[mHMMbayes]{mHMM_cont}.}

\item{mcmc_burn_in}{Int. number of burn-in samples for the MCMC algorithm. Defaults to 500. See mcmc parameter in \link[mHMMbayes]{mHMM_cont}.}

\item{show_progress}{Boolean. Should progress of MCMC algorithm be displayed? Defaults to TRUE.}
}
\value{
An mHMM_cont object containing posterior distributions for each of the parameters.
}
\description{
Run an mHMM on a simulated sleep dataset
}
