% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Simulate a sleep dataset for use in an mHMM model}
\usage{
simulate_dataset(n, n_t, zeta, Q, seed)
}
\arguments{
\item{n}{int. number of subjects.}

\item{n_t}{int. number of observations for each subject.}

\item{zeta}{float. between-subject variance of the emission distributions.}

\item{Q}{float. between-subject variance of the transition probability matrix.}

\item{seed}{int. random seed used to generate the dataset.}
}
\value{
list containing states and emission distribution observed data. See \link[mHMMbayes]{sim_mHMM}.
}
\description{
Simulate a sleep dataset for use in an mHMM model
}
