% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trace_plot}
\alias{trace_plot}
\title{Make a trace plot of the data}
\usage{
trace_plot(x, ...)
}
\arguments{
\item{x}{an mHMM_cont object}

\item{param}{string. Which of the parameters should be plotted? Must be one of 'emiss_mu_bar', 'gamma_int_bar', 'emiss_var_bar' or 'emiss_varmu_bar'}

\item{var}{int. Which emission distribution should be plotted? Must be an integer equal to or less than the total number of emission distributions. Index of variables is the same as the order in which they appear in the input data given to the mHMM.}
}
\value{
XX
}
\description{
Make a trace plot of the data
}
