% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_posterior_means}
\alias{plot_posterior_means}
\title{Plot histograms of between-subject means}
\usage{
plot_posterior_means(x, ...)
}
\arguments{
\item{x}{an mHMM_cont object}

\item{param}{string. Which of the parameters should be plotted? Must be one of 'between_means', 'tpm_intercepts', 'residual_var' or 'between_var'}

\item{var}{int. Which emission distribution should be plotted? Must be an integer equal to or less than the total number of emission distributions.}

\item{ground_truth}{numeric. Vector of ground-truth values for the parameters equal to the number of emission distributions. Will be added as a red dotted line.}
}
\value{
plots the between-subject means
}
\description{
Plot histograms of between-subject means
}
