% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Simulate a sleep dataset for use in an mHMM model}
\usage{
simulate_dataset(n, n_t, zeta, Q)
}
\arguments{
\item{Q}{}
}
\value{

}
\description{
Simulate a sleep dataset for use in an mHMM model
}
